Update list cmd to use print_error and exit
Since we are now using `exit`, we have to capture the raised error otherwise rspec will exit successfully once it encounters it. Based on what ronin CLI does, we have to assert the `print_error` call to validate that we display the right error message since we can't check the stderr output.